==> notebooks/00_data_quality_and_validation.ipynb <==
import geopandas as gpd
import matplotlib.pyplot as plt

# 1. Load the Berlin district boundaries 
districts = gpd.read_file("../maps/bezirksgrenzen.geojson")

# Inspect the columns to know names
print(districts.columns)

# Inspect
print(districts['Gemeinde_name'].unique())



# Filter for MITTE
mitte = districts[districts['Gemeinde_name'] == "Mitte"]


# Save Mitte-only district
mitte.to_file("../maps/mitte.geojson", driver="GeoJSON")



mitte

mitte = gpd.read_file("../maps/mitte.geojson")
print(mitte)
print(mitte.geometry.iloc[0])


print(type(mitte.geometry.iloc[0]))


import geopandas as gpd
from shapely.geometry import Polygon, MultiPolygon
import osmnx as ox

# 1️⃣ Read your file
mitte = gpd.read_file("../maps/mitte.geojson")

# 2️⃣ Get the first geometry
mitte_poly = mitte.geometry.iloc[0]

# 3️⃣ Check what’s inside (if MultiPolygon)
if isinstance(mitte_poly, MultiPolygon):
    for i, geom in enumerate(mitte_poly.geoms):
        print(i, geom.geom_type)


from shapely.geometry import Polygon, MultiPolygon

valid_polys = [g for g in mitte_poly.geoms if isinstance(g, (Polygon, MultiPolygon))]

# Rebuild MultiPolygon with only valid polygons
mitte_poly = MultiPolygon(valid_polys)
