==> notebooks/01_study_area_definition.ipynb <==
import os

os.chdir(r"C:\Users\romak\Desktop\Urban_Technology_Berlin_Mitte")
os.getcwd()


import geopandas as gpd
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt


mitte = gpd.read_file("maps/mitte.geojson")
buildings = gpd.read_file("data/processed/buildings_mitte.geojson")
public_spaces = gpd.read_file("data/processed/mitte_public_spaces.geojson")


print("Mitte:", mitte.shape)
print("Buildings:", buildings.shape)
print("Public spaces:", public_spaces.shape)


print("Mitte CRS:", mitte.crs)
print("Buildings CRS:", buildings.crs)
print("Public spaces CRS:", public_spaces.crs)


mitte = mitte.to_crs(epsg=25833)


print("Mitte CRS:", mitte.crs)
print("Buildings CRS:", buildings.crs)
print("Public spaces CRS:", public_spaces.crs)


buildings.total_bounds


buildings["bezeich"].value_counts()


buildings_clean = buildings[buildings["bezeich"] == "AX_Gebaeude"].copy()
buildings_clean.shape


buildings_clean.geom_type.value_counts()


buildings_clean.is_valid.value_counts()


print("Invalid buildings:", (~buildings.is_valid).sum())
print("Invalid public spaces:", (~public_spaces.is_valid).sum())


print("Invalid buildings:", (~buildings.is_valid).sum())
print("Invalid public spaces:", (~public_spaces.is_valid).sum())


buildings.geometry.type.value_counts()


public_spaces.geometry.type.value_counts()


buildings["height_m"].describe()


buildings.columns



buildings_with_height = gpd.read_file(
    "data/processed/buildings_with_height.geojson"
)



buildings_with_height.columns


buildings_with_height["height_m"].describe()
