==> notebooks/05_solar_geometry_calculation.ipynb <==
import os
os.chdir(r"C:\Users\romak\Desktop\Urban_Technology_Berlin_Mitte")
os.getcwd()


import pandas as pd
import pvlib


latitude = 52.5200
longitude = 13.4050

location = pvlib.location.Location(latitude, longitude, tz="Europe/Berlin")
location


time = pd.DatetimeIndex(
    ["2025-12-21 12:00"],
    tz="Europe/Berlin"
)

time


sun_pos = location.get_solarposition(time)

sun_pos


altitude = sun_pos["apparent_elevation"].iloc[0]
azimuth = sun_pos["azimuth"].iloc[0]

altitude, azimuth
